ClientInstance / Client Handler now send/receive correct number of messages
After many experiments, found that JAXB retains previously sent data inside the marshaller, and by recreating the marshaller and/or context each time data needs to be sent, data is only sent once. Much experimental code will need to be removed, because I originally thought the problem was with concurrency, and added a large number of safeties to prevent concurrency problems that appear now to be unnecessary as that was not the problem.